requirements.txt
Flask
gunicorn
Pillow
numpy
opencv-python-headless
